refactor: collect dynamic scaling of geoJson objects into Map.js Also, remove strange resize detector as it seems to do nothing
