+++ app/component/map/Map.js
@@ -1,8 +1,5 @@
 import PropTypes from 'prop-types';
 import React from 'react';
-
-import elementResizeDetectorMaker from 'element-resize-detector';
-
 import LeafletMap from 'react-leaflet/es/Map';
 import TileLayer from 'react-leaflet/es/TileLayer';
 import AttributionControl from 'react-leaflet/es/AttributionControl';
@@ -14,7 +11,11 @@ import isString from 'lodash/isString';
 import isEmpty from 'lodash/isEmpty';
 // Webpack handles this by bundling it with the other css files
 import 'leaflet/dist/leaflet.css';
-
+import VehicleMarkerContainer from './VehicleMarkerContainer';
+import {
+  startRealTimeClient,
+  stopRealTimeClient,
+} from '../../action/realTimeClientAction';
 import PositionMarker from './PositionMarker';
 import VectorTileLayerContainer from './tile-layer/VectorTileLayerContainer';
 import { boundWithMinimumArea } from '../../util/geo-utils';
@@ -25,8 +26,28 @@ import events from '../../util/events';
 import GeoJSON from './GeoJSON';
 
 const zoomOutText = `<svg class="icon"><use xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="#icon-icon_minus"/></svg>`;
-
 const zoomInText = `<svg class="icon"><use xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="#icon-icon_plus"/></svg>`;
+
+/* foo-eslint-disable react/sort-comp */
+
+const startClient = context => {
+  const { realTime } = context.config;
+  let agency;
+  /* handle multiple feedid case */
+  context.config.feedIds.forEach(ag => {
+    if (!agency && realTime[ag]) {
+      agency = ag;
+    }
+  });
+  const source = agency && realTime[agency];
+  if (source && source.active) {
+    const config = {
+      ...source,
+      agency,
+    };
+    context.executeAction(startRealTimeClient, config);
+  }
+};
 
 export default class Map extends React.Component {
   static propTypes = {
@@ -66,14 +87,42 @@ export default class Map extends React.Component {
 
   static contextTypes = {
     executeAction: PropTypes.func.isRequired,
+    getStore: PropTypes.func,
     config: PropTypes.object.isRequired,
   };
 
+  constructor(props) {
+    super(props);
+    this.state = { zoom: 14 };
+  }
+
+  updateZoom = () => {
+    // eslint-disable-next-line no-underscore-dangle
+    const zoom = this.map?.leafletElement?._zoom || this.props.zoom || 16;
+    if (zoom !== this.state.zoom) {
+      this.setState({ zoom });
+    }
+  };
+
   componentDidMount() {
-    /* eslint-disable no-underscore-dangle */
-    if (this.map) {
-      this.erd = elementResizeDetectorMaker({ strategy: 'scroll' });
-      this.erd.listenTo(this.map.leafletElement._container, this.resizeMap);
+    if (this.props.mapLayers.vehicles) {
+      startClient(this.context);
+    }
+    this.updateZoom();
+  }
+
+  // eslint-disable-next-line camelcase
+  UNSAFE_componentWillReceiveProps(newProps) {
+    this.updateZoom();
+    if (newProps.mapLayers.vehicles) {
+      if (!this.props.mapLayers.vehicles) {
+        startClient(this.context);
+      }
+    } else if (this.props.mapLayers.vehicles) {
+      const { client } = this.context.getStore('RealTimeInformationStore');
+      if (client) {
+        this.context.executeAction(stopRealTimeClient, client);
+      }
     }
   }
 
@@ -88,20 +137,17 @@ export default class Map extends React.Component {
   }
 
   componentWillUnmount() {
-    if (this.erd) {
-      this.erd.removeListener(
-        this.map.leafletElement._container,
-        this.resizeMap,
-      );
+    const { client } = this.context.getStore('RealTimeInformationStore');
+    if (client) {
+      this.context.executeAction(stopRealTimeClient, client);
     }
   }
 
   onPopupopen = () => events.emit('popupOpened');
 
-  resizeMap = () => {
-    if (this.map) {
-      this.map.leafletElement.invalidateSize(false);
-    }
+  zoomEnd = () => {
+    this.props.leafletEvents?.onZoomEnd?.(); // pass event to parent
+    this.updateZoom();
   };
 
   render() {
@@ -134,12 +180,6 @@ export default class Map extends React.Component {
       }
     }
 
-    if (!this.erd && this.map) {
-      this.erd = elementResizeDetectorMaker({ strategy: 'scroll' });
-      /* eslint-disable no-underscore-dangle */
-      this.erd.listenTo(this.map.leafletElement._container, this.resizeMap);
-    }
-
     if (naviProps.bounds || (naviProps.center && naviProps.zoom)) {
       this.ready = true;
     }
@@ -169,12 +209,21 @@ export default class Map extends React.Component {
       />,
     ]);
 
+    if (this.props.mapLayers.vehicles) {
+      const useLargeIcon = this.state.zoom >= config.stopsMinZoom;
+      leafletObjNew.push(
+        <VehicleMarkerContainer
+          key="vehicles"
+          useLargeIcon={useLargeIcon}
+          ignoreMode
+        />,
+      );
+    }
+
     let attribution = get(config, 'map.attribution');
     if (!isString(attribution) || isEmpty(attribution)) {
       attribution = false;
     }
-
-    const mapZoom = this.map?.leafletElement.getZoom() || zoom || 14;
 
     if (geoJson) {
       Object.keys(geoJson)
@@ -189,13 +238,19 @@ export default class Map extends React.Component {
             <GeoJSON
               key={key.concat(i)}
               data={geoJson[key].data}
-              geoJsonZoomLevel={mapZoom}
+              geoJsonZoomLevel={this.state.zoom}
               locationPopup={locationPopup}
               onSelectLocation={onSelectLocation}
             />,
           );
         });
     }
+
+    const leafletEvents = {
+      ...this.props.leafletEvents,
+      onZoomend: this.zoomEnd,
+    };
+
     return (
       <div aria-hidden="true">
         <span
@@ -208,7 +263,7 @@ export default class Map extends React.Component {
         </span>
         <LeafletMap
           {...naviProps}
-          className={`z${mapZoom}`}
+          className={`z${this.state.zoom}`}
           keyboard={false}
           ref={el => {
             this.map = el;
@@ -222,7 +277,7 @@ export default class Map extends React.Component {
           attributionControl={false}
           animate={this.props.animate}
           boundsOptions={boundsOptions}
-          {...this.props.leafletEvents}
+          {...leafletEvents}
           onPopupopen={this.onPopupopen}
           closePopupOnClick={false}
         >
