+++ app/component/map/MapWithTracking.js
@@ -12,12 +12,6 @@ import MapContainer from './MapContainer';
 import ToggleMapTracking from '../ToggleMapTracking';
 import { isBrowser } from '../../util/browser';
 import PositionStore from '../../store/PositionStore';
-import VehicleMarkerContainer from './VehicleMarkerContainer';
-import {
-  startRealTimeClient,
-  stopRealTimeClient,
-} from '../../action/realTimeClientAction';
-import { addAnalyticsEvent } from '../../util/analyticsUtils';
 import { mapLayerShape } from '../../store/MapLayerStore';
 
 const onlyUpdateCoordChanges = onlyUpdateForKeys([
@@ -32,28 +26,6 @@ const onlyUpdateCoordChanges = onlyUpdateForKeys([
 ]);
 
 const MapCont = onlyUpdateCoordChanges(MapContainer);
-
-/* stop yet another eslint madness */
-/* eslint-disable react/sort-comp */
-
-const startClient = context => {
-  const { realTime } = context.config;
-  let agency;
-  /* handle multiple feedid case */
-  context.config.feedIds.forEach(ag => {
-    if (!agency && realTime[ag]) {
-      agency = ag;
-    }
-  });
-  const source = agency && realTime[agency];
-  if (source && source.active) {
-    const config = {
-      ...source,
-      agency,
-    };
-    context.executeAction(startRealTimeClient, config);
-  }
-};
 
 class MapWithTrackingStateHandler extends React.Component {
   static propTypes = {
@@ -101,15 +73,6 @@ class MapWithTrackingStateHandler extends React.Component {
     if (!isBrowser) {
       return;
     }
-
-    if (this.props.mapLayers.vehicles) {
-      startClient(this.context);
-      const currentZoom = // eslint-disable-next-line no-underscore-dangle
-        this.mapElement?.leafletElement?._zoom || this.props.zoom || 16;
-      if (currentZoom !== this.state.vehicleZoom) {
-        this.setState({ vehicleZoom: currentZoom });
-      }
-    }
     if (this.props.setMWTRef) {
       this.props.setMWTRef(this);
     }
@@ -123,33 +86,8 @@ class MapWithTrackingStateHandler extends React.Component {
     } else if (newProps.mapTracking === false && this.state.mapTracking) {
       newState = { mapTracking: false };
     }
-    if (newProps.mapLayers.vehicles) {
-      const currentZoom = // eslint-disable-next-line no-underscore-dangle
-        this.mapElement?.leafletElement?._zoom || newProps.zoom || 16;
-      if (currentZoom !== this.state.vehicleZoom) {
-        newState = newState || {};
-        newState.vehicleZoom = currentZoom;
-      }
-    }
     if (newState) {
       this.setState(newState);
-    }
-    if (newProps.mapLayers.vehicles) {
-      if (!this.props.mapLayers.vehicles) {
-        startClient(this.context);
-      }
-    } else if (this.props.mapLayers.vehicles) {
-      const { client } = this.context.getStore('RealTimeInformationStore');
-      if (client) {
-        this.context.executeAction(stopRealTimeClient, client);
-      }
-    }
-  }
-
-  componentWillUnmount() {
-    const { client } = this.context.getStore('RealTimeInformationStore');
-    if (client) {
-      this.context.executeAction(stopRealTimeClient, client);
     }
   }
 
@@ -172,11 +110,6 @@ class MapWithTrackingStateHandler extends React.Component {
     if (this.props.onMapTracking) {
       this.props.onMapTracking();
     }
-    addAnalyticsEvent({
-      category: 'Map',
-      action: 'ReCenterToMyGeolocation',
-      name: null,
-    });
   };
 
   disableMapTracking = () => {
@@ -203,11 +136,6 @@ class MapWithTrackingStateHandler extends React.Component {
       this.props.onEndNavigation(this.mapElement);
     }
     this.navigated = true;
-    // eslint-disable-next-line no-underscore-dangle
-    const zoom = this.mapElement?.leafletElement?._zoom;
-    if (this.props.mapLayers.vehicles && zoom !== this.state.vehicleZoom) {
-      this.setState({ vehicleZoom: zoom });
-    }
   };
 
   render() {
@@ -224,20 +152,6 @@ class MapWithTrackingStateHandler extends React.Component {
       ...rest
     } = this.props;
     const { config } = this.context;
-    const leafletObjs = [];
-    if (this.props.leafletObjs) {
-      leafletObjs.push(...this.props.leafletObjs);
-    }
-    if (this.props.mapLayers.vehicles) {
-      const useLargeIcon = this.state.vehicleZoom >= config.stopsMinZoom;
-      leafletObjs.push(
-        <VehicleMarkerContainer
-          key="vehicles"
-          useLargeIcon={useLargeIcon}
-          ignoreMode
-        />,
-      );
-    }
 
     let btnClassName = 'map-with-tracking-buttons'; // DT-3470
     if (config.map.showZoomControl) {
@@ -311,7 +225,6 @@ class MapWithTrackingStateHandler extends React.Component {
         }}
         {...this.naviProps}
         {...rest}
-        leafletObjs={leafletObjs}
         mapRef={this.setMapElementRef}
         bottomButtons={
           <div className={btnClassName}>
@@ -348,15 +261,7 @@ class MapWithTrackingStateHandler extends React.Component {
 MapWithTrackingStateHandler.contextTypes = {
   executeAction: PropTypes.func,
   getStore: PropTypes.func,
-  config: PropTypes.shape({
-    realTime: PropTypes.object.isRequired,
-    feedIds: PropTypes.array.isRequired,
-    stopsMinZoom: PropTypes.number.isRequired,
-    geoJson: PropTypes.shape({
-      layers: PropTypes.array,
-      layerConfigUrl: PropTypes.string,
-    }),
-  }).isRequired,
+  config: PropTypes.object.isRequired,
 };
 
 const MapWithTracking = connectToStores(
