+++ package.json
@@ -146,7 +146,6 @@
     "connect-redis": "5.0.0",
     "cookie-parser": "1.4.5",
     "debug": "4.1.1",
-    "element-resize-detector": "1.2.0",
     "esm": "3.2.25",
     "express": "4.17.1",
     "express-http-proxy": "1.6.2",
